refactor: make logging proxies builder-native
